# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Docs",
        "version": "1.0.0"
    },
    "basePath": "/api",
    "schemes": [
        "https",
        "http"
    ],
    "paths": {
...
        "/focuses": {
            "get": {
                "tags": [
                    "Focuses"
                ],
                "summary": "Get list of focuses",
                "parameters": [
                    {
                        "name": "sort",
                        "in": "query",
                        "description": "Direction of sorting (asc, desc) Default: asc",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Count of records per page. Default: 10",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset of records. Default: 0",
                        "required": false,
                        "type": "integer"
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Status. (Draft - 1, Review - 2, Published - 3)",
                        "required": false,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of focuses",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/OffsetLimitData"
                                },
                                {
                                    "properties": {
                                        "data": {
                                            "items": {
                                                "$ref": "#/definitions/Focus"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "403": {
                        "description": "Unauthorized action"
                    }
                },
                "security": [
                    {
                        "apiToken": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Focuses"
                ],
                "summary": "Create new Focus",
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Focus Body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/FocusBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "properties": {
                                "focus": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Focus"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "403": {
                        "description": "Unauthorized action"
                    },
                    "400": {
                        "description": "Validation Error",
                        "schema": {
                            "$ref": "#/definitions/Validation Error"
                        }
                    }
                },
                "security": [
                    {
                        "apiToken": []
                    }
                ]
            }
        },
        "/focuses/{id}": {
            "get": {
                "tags": [
                    "Focuses"
                ],
                "summary": "Get focus data",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Focus id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success operation",
                        "schema": {
                            "properties": {
                                "focus": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Focus"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "403": {
                        "description": "Unauthorized action"
                    }
                },
                "security": [
                    {
                        "apiToken": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Focuses"
                ],
                "summary": "Update focus. All properties is required.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Focus id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Focus body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/FocusBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Focus was successfully updated",
                        "schema": {
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Focus"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "403": {
                        "description": "Unauthorized action"
                    },
                    "400": {
                        "description": "Validation Error",
                        "schema": {
                            "$ref": "#/definitions/Validation Error"
                        }
                    }
                },
                "security": [
                    {
                        "apiToken": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Focuses"
                ],
                "summary": "Delete focus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Focus id",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Focus was successfully deleted"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "403": {
                        "description": "Unauthorized action"
                    }
                },
                "security": [
                    {
                        "apiToken": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Focuses"
                ],
                "summary": "Update focus properties",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Focus id",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Focus body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/FocusBody"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Focus was successfully updated",
                        "schema": {
                            "properties": {
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Focus"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "403": {
                        "description": "Unauthorized action"
                    },
                    "400": {
                        "description": "Validation Error",
                        "schema": {
                            "$ref": "#/definitions/Validation Error"
                        }
                    }
                },
                "security": [
                    {
                        "apiToken": []
                    }
                ]
            }
        },
...
        "/focuses/{id}/reset": {
            "post": {
                "tags": [
                    "Focuses"
                ],
                "summary": "Reset progress of Focus",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Focus ID",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Focus data"
                    },
                    "401": {
                        "description": "Unauthorized user"
                    },
                    "403": {
                        "description": "Unauthorized action"
                    }
                },
                "security": [
                    {
                        "apiToken": []
                    }
                ]
            }
        },
...
    },
    "definitions": {
...
        "OffsetLimitData": {
            "properties": {
                "offset": {
                    "type": "integer",
                    "default": 0
                },
                "limit": {
                    "type": "integer",
                    "default": 10
                }
            }
        },
...
        "CreatedUpdatedDates": {
            "properties": {
                "created_at": {
                    "description": "Format(PHP): Y:m:d H:i:s O",
                    "type": "string"
                },
                "updated_at": {
                    "description": "Format(PHP): Y:m:d H:i:s O",
                    "type": "string"
                }
            }
        },
...
        "Focus": {
            "allOf": [
                {
                    "properties": {
                        "id": {
                            "type": "integer"
                        },
                        "title": {
                            "type": "string"
                        },
                        "slug": {
                            "type": "string"
                        },
                        "status": {
                            "description": "Focus status, available statuses: 1 - Draft | 2 - Review | 3 - Published",
                            "type": "integer"
                        },
                        "accent_color": {
                            "type": "string"
                        },
                        "image": {
                            "description": "Directed link to image",
                            "type": "string"
                        },
                        "count_sessions": {
                            "type": "integer"
                        }
                    }
                },
                {
                    "$ref": "#/definitions/CreatedUpdatedDates"
                }
            ]
        },
...
        "FocusBody": {
            "required": [
                "title",
                "slug",
                "status"
            ],
            "properties": {
                "program_id": {
                    "description": "The Identificator of the existing Program",
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "status": {
                    "description": "available statuses: 1 - Draft | 2 - Review | 3 - Published",
                    "type": "integer"
                },
                "accent_color": {
                    "description": "In HEX format Example: #aaff00",
                    "type": "string"
                },
                "image_url": {
                    "description": "Link to image",
                    "type": "string"
                },
                "video_overview": {
                    "description": "Link to video",
                    "type": "string"
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "apiToken": {
            "type": "apiKey",
            "description": "Api token",
            "name": "Authorization",
            "in": "header"
        }
    }
}